0037-pyarrow/requirements.txt
loguru
pandas
pyarrow
pyyaml
s3fs
tqdm